orchestration/pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "acis-orchestration"
version = "1.0.0"

[project.optional-dependencies]
dev = ["dagster-webserver", "pytest"]
